+++ src/controllers/users.js
@@ -1,13 +1,13 @@
-const ErrorResponse = require('../utils/errorResponse');
+// const ErrorResponse = require('../utils/errorResponse');
 
-/* This is an example of a router controller */
-exports.getUser = async (req, res, next) => {
-  /* const user = await User.findOne({});
-  if (!user) {
-    // This is an example of an error
-    return next(new ErrorResponse('User not found', 404));
-  }
+// /* This is an example of a router controller */
+// exports.getUser = async (req, res, next) => {
+//   /* const user = await User.findOne({});
+//   if (!user) {
+//     // This is an example of an error
+//     return next(new ErrorResponse('User not found', 404));
+//   }
 
-  // This is an example of a successful response
-  res.json({ success: true, data: user }); */
-};
+//   // This is an example of a successful response
+//   res.json({ success: true, data: user }); */
+// };
