+++ src/routes/index.js
@@ -1,7 +1,8 @@
 const router = require('express').Router();
 
 const usersRoutes = require('./users');
+const authRoutes = require('./auth');
 
 router.use('/users', usersRoutes);
-
+router.use('/auth', authRoutes);
 module.exports = router;
